Update Gravityless tricks.
Added Gravityless frigate front to back and back to freight lift nsj (Wave beam is not check for when going from back to front so this may softlock/bug).
Changed Remove Gravity Requirements to not require SJ for tower chamber gravityless. Note that Tower Chamber NSJ trick is different as that one puts it in logic with gravity and bombs if Remove Gravity Requirements is active.

Remove Gravity Requirements trick overrules all but gravityless frigate NSJ, reason for that being is that the NSJ trick is harder than the SJ one so the NSJ trick if enabled will overrule the SJ one.